Workaround for black screenshots due to z-order change
The screenshotApplications method in WindowManagerService.java determines
the minz and maxz layers to screenshot. A black screenshot is taken because
the z-order of the target window(s) has changed by the time the native
screenshot function is called.

It appears the target window is usually ends up a small offset above the
maxLayerZ. If there are no windows in the z-range then increase the
maxLayerZ so that the target window might be captured.